tools/requirements.txt
Pillow
ffmpeg-python
pyyaml
